+++ app/src/main/java/org/openziti/mobile/ZitiMobileEdgeApp.kt
@@ -9,6 +9,7 @@ import android.content.Intent
 import android.util.Log
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
 import org.openziti.mobile.debug.DebugInfo
+import org.openziti.mobile.model.TunnelModel
 
 import org.openziti.tunnel.Tunnel
